Removing un-necessary code
